Added comparaison script. Divise the file in multiple smaller files for clarity and modularity
